requirements.txt
# Background Removal Service - Python Dependencies
#
# Install with: pip install -r requirements.txt

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Configuration
pydantic-settings>=2.0.0

# Image processing
pillow>=10.0.0

# Background removal models
# rembg requires Python >=3.11, <3.14 (per their docs)
# Force modern numba for Python 3.11+ compatibility (pymatting constraint is too loose)
numba>=0.59.0
rembg[cpu]>=2.0.72
withoutbg>=1.0.0

# HTTP client for batch testing
requests>=2.31.0
